# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "RCI-BACKEND API DOCUMENTATION",
        "description": "API DOCUMENTATION",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://talararci.thonymarckdev.online"
        }
    ],
    "paths": {
...
        "/api/obtenerInfoAdmins": {
            "get": {
                "tags": [
                    "DASHBOARD CONTROLLER"
                ],
                "summary": "Obtener estadísticas de usuarios",
                "description": "Permite obtener estadísticas sobre los usuarios, incluyendo el número de usuarios activos, inactivos y el total de usuarios.",
                "operationId": "obtenerInfoAdmins",
                "responses": {
                    "200": {
                        "description": "Estadísticas de usuarios obtenidas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "stats": {
                                            "properties": {
                                                "active_users": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "inactive_users": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "total_users": {
                                                    "type": "integer",
                                                    "example": 15
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User statistics retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error interno del servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/logs": {
            "get": {
                "tags": [
                    "DASHBOARD CONTROLLER"
                ],
                "summary": "Obtener registros de logs",
                "description": "Permite obtener los registros de logs filtrados por fecha, usuario, acción y rol. El acceso a los logs está restringido según el rol del usuario autenticado.",
                "operationId": "getLogs",
                "parameters": [
                    {
                        "name": "fecha_desde",
                        "in": "query",
                        "description": "Fecha de inicio para filtrar los logs (formato YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "fecha_hasta",
                        "in": "query",
                        "description": "Fecha de fin para filtrar los logs (formato YYYY-MM-DD)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "usuario",
                        "in": "query",
                        "description": "Nombre del usuario para filtrar los logs",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "accion",
                        "in": "query",
                        "description": "Acción para filtrar los logs",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rol",
                        "in": "query",
                        "description": "Rol para filtrar los logs",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "admin",
                                "superadmin"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de logs y usuarios filtrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "logs": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "idUsuario": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "accion": {
                                                        "type": "string",
                                                        "example": "login"
                                                    },
                                                    "fecha": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2023-10-01T12:00:00Z"
                                                    },
                                                    "nombres": {
                                                        "type": "string",
                                                        "example": "Juan Pérez"
                                                    },
                                                    "rol": {
                                                        "type": "string",
                                                        "example": "admin"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "usuarios": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "idUsuario": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nombres": {
                                                        "type": "string",
                                                        "example": "Juan Pérez"
                                                    },
                                                    "rol": {
                                                        "type": "string",
                                                        "example": "admin"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "No autorizado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error interno del servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Usar un token JWT en el encabezado Authorization como Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}